DSS-3650 : Provide more meaningful exception message on failed augmentation of a CMS signature with corrupted content
